requirements.txt
charset-normalizer==3.4.1
ffmpeg-python==0.2.0
future==1.0.0
mysql-connector-python==9.2.0
